@@ -9,10 +9,10 @@ using Xunit.Abstractions;
 
 namespace Trakx.Shrimpy.ApiClient.Tests.Integration
 {
-    public class AccountsClientTests : ShrimpyClientTestsBase
+    public sealed class AccountsClientTests : ShrimpyClientTestsBase
     {
         private readonly IAccountsClient _accountsClient;
-        public AccountsClientTests(ITestOutputHelper output) : base(output)
+        public AccountsClientTests(ShrimpyApiFixture apiFixture, ITestOutputHelper output) : base(apiFixture, output)
         {
             _accountsClient = ServiceProvider.GetRequiredService<IAccountsClient>();
         }
