@@ -1,9 +1,13 @@
 ﻿
+using System.Net.Http;
+
 namespace Trakx.Shrimpy.ApiClient
 {
     internal partial class AccountsClient
     {
-        partial void PrepareRequest(System.Net.Http.HttpClient client, System.Net.Http.HttpRequestMessage request, string url)
+#pragma warning disable S1172 // Unused method parameters should be removed
+        partial void PrepareRequest(HttpClient client, HttpRequestMessage request, string url)
+#pragma warning restore S1172 // Unused method parameters should be removed
         {
             CredentialProvider.AddCredentials(request);
         }
