@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Trakx.Shrimpy.ApiClient.Utils;
+using Trakx.Utils.Apis;
 
 namespace Trakx.Shrimpy.ApiClient
 {
@@ -12,7 +12,7 @@ namespace Trakx.Shrimpy.ApiClient
         public ClientConfigurator(IServiceProvider serviceProvider)
         {
             _serviceProvider = serviceProvider;
-            ApiConfiguration = serviceProvider.GetService<IOptions<ShrimpyApiConfiguration>>().Value;
+            ApiConfiguration = serviceProvider.GetService<IOptions<ShrimpyApiConfiguration>>()!.Value;
         }
 
         public ShrimpyApiConfiguration ApiConfiguration { get; }
@@ -25,7 +25,7 @@ namespace Trakx.Shrimpy.ApiClient
                 //case nameof(ExchangesClient):
                     return new NoCredentialsProvider();
                 default:
-                    return _serviceProvider.GetService<ICredentialsProvider>();
+                    return _serviceProvider.GetService<ICredentialsProvider>()!;
             }
         }
     }
