@@ -6,6 +6,8 @@ using System.Text;
 using System.Threading;
 using Microsoft.Extensions.Options;
 using Serilog;
+using Trakx.Utils.Apis;
+using Trakx.Utils.DateTimeHelpers;
 
 namespace Trakx.Shrimpy.ApiClient.Utils
 {
@@ -38,7 +40,7 @@ namespace Trakx.Shrimpy.ApiClient.Utils
         /// <inheritdoc />
         public void AddCredentials(HttpRequestMessage msg)
         {
-            var path = msg.RequestUri.AbsolutePath;
+            var path = msg.RequestUri!.AbsolutePath;
             var method = msg.Method.Method.ToUpperInvariant();
             var nonce = GetNonce();
             var body = msg.Content?.ReadAsStringAsync().GetAwaiter().GetResult() ?? string.Empty;
